refactor page and create custom hooks
